requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bs4==0.0.2
cashews==7.3.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
distlib==0.3.8
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
nb-cli==1.4.2
nonebot-adapter-onebot==2.4.5
nonebot-plugin-guild-patch==0.2.3
nonebot2==2.3.3
noneprompt==0.1.9
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
pyfiglet==1.0.2
Pygments==2.18.0
pygtrie==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
text-unidecode==1.3
tomlkit==0.13.2
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.5
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.13.1
